tests: Test Provider respects INCUS_REMOTE
When INCUS_REMOTE is set then the Test Provider loads the user's local
configuration and sets the provided INCUS_REMOTE as the default remote
to execute the tests.

This is very handy when tests are running locally, e.g. virtual machine
specific tests which are not working in GitHub Actions.

Signed-off-by: Fabian Mettler <dev@maveonair.com>